move notification default layout to src/core
