+++ src/core/notification/layouts/default/index.ts
