+++ src/core/notification/layouts/default/notification-layout-default.element.ts
@@ -2,12 +2,9 @@ import { html, LitElement } from 'lit';
 import { customElement, property } from 'lit/decorators.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { UUITextStyles } from '@umbraco-ui/uui-css';
-import type { UmbNotificationHandler } from '../..';
+import type { UmbNotificationDefaultData, UmbNotificationHandler } from '@umbraco-cms/notification';
 
-export interface UmbNotificationDefaultData {
-	message: string;
-	headline?: string;
-}
+export type { UmbNotificationDefaultData };
 
 @customElement('umb-notification-layout-default')
 export class UmbNotificationLayoutDefaultElement extends LitElement {
@@ -25,5 +22,11 @@ export class UmbNotificationLayoutDefaultElement extends LitElement {
 				<div id="message">${this.data.message}</div>
 			</uui-toast-notification-layout>
 		`;
+	}
+}
+
+declare global {
+	interface HTMLElementTagNameMap {
+		'umb-notification-layout-default': UmbNotificationLayoutDefaultElement;
 	}
 }
