+++ src/core/notification/stories/story-notification-default-example.element.ts
@@ -1,12 +1,11 @@
 import { html } from 'lit';
 import { customElement } from 'lit/decorators.js';
-import { UmbNotificationDefaultData } from '../layouts/default';
 import {
 	UmbNotificationColor,
 	UmbNotificationOptions,
 	UmbNotificationContext,
 	UMB_NOTIFICATION_CONTEXT_TOKEN,
-} from '..';
+} from '@umbraco-cms/notification';
 import { UmbLitElement } from '@umbraco-cms/element';
 
 @customElement('story-notification-default-example')
@@ -22,7 +21,7 @@ export class StoryNotificationDefaultExampleElement extends UmbLitElement {
 	}
 
 	private _handleNotification = (color: UmbNotificationColor) => {
-		const options: UmbNotificationOptions<UmbNotificationDefaultData> = {
+		const options: UmbNotificationOptions = {
 			data: {
 				headline: 'Headline',
 				message: 'Lorem ipsum dolor sit amet, consectetur adipiscing elit',
@@ -52,5 +51,11 @@ export class StoryNotificationDefaultExampleElement extends UmbLitElement {
 
 			<umb-backoffice-notification-container></umb-backoffice-notification-container>
 		`;
+	}
+}
+
+declare global {
+	interface HTMLElementTagNameMap {
+		'story-notification-default-example': StoryNotificationDefaultExampleElement;
 	}
 }
